backend/requirements.txt
Django==4.0.2
Pillow==9.0.0
requests==2.27.1
djangorestframework==3.13.1